refactor(agentic): modularize synapse-agentic single-file crate
