+++ synapse-agentic/src/lib.rs
@@ -1,898 +1,35 @@
+pub mod context;
+pub mod providers;
+pub mod router;
+pub mod graph;
+
 pub mod prelude {
     pub use async_trait::async_trait;
     pub use serde::{Deserialize, Serialize};
     pub use serde_json::Value;
 
-    use std::sync::Arc;
-    use tokio::sync::mpsc;
+    pub use crate::context::{
+        CompactionConfig, ContextOverflowRisk, DecisionContext, ExplicitPlanner, LLMSummarizer,
+        Message, MessageChunk, MessageRole, PlannedTask, SessionContext, SimpleTokenEstimator,
+        SummarizationStrategy, TaskStatus, TokenCounter,
+    };
 
-    #[derive(Debug, Clone, Serialize, Deserialize)]
-    pub struct DecisionContext {
-        pub query: String,
-        pub summary: Option<String>,
-        pub metadata: std::collections::HashMap<String, String>,
-        pub data: Option<Value>,
-    }
-    impl DecisionContext {
-        pub fn new(q: &str) -> Self {
-            Self {
-                query: q.to_string(),
-                summary: None,
-                metadata: std::collections::HashMap::new(),
-                data: None,
-            }
-        }
-        pub fn with_metadata(mut self, k: &str, v: String) -> Self {
-            self.metadata.insert(k.to_string(), v);
-            self
-        }
-        pub fn with_summary(mut self, s: impl Into<String>) -> Self {
-            self.summary = Some(s.into());
-            self
-        }
-        pub fn with_data(mut self, d: Value) -> Self {
-            self.data = Some(d);
-            self
-        }
-    }
+    pub use crate::providers::{
+        GeminiProvider, GroqProvider, LLMProvider, MinimaxProvider, Provider,
+    };
 
-    #[async_trait]
-    pub trait LLMProvider: Send + Sync + std::fmt::Debug {
-        fn name(&self) -> &str;
-        fn cost_per_1k_tokens(&self) -> f64;
-        async fn generate(&self, prompt: &str) -> anyhow::Result<String>;
-    }
-
-    pub trait Provider: LLMProvider {}
-    impl<T: LLMProvider> Provider for T {}
-
-    #[derive(Debug, Clone)]
-    pub struct GeminiProvider {
-        api_key: String,
-        model: String,
-    }
-    impl GeminiProvider {
-        pub fn new(api_key: String, model: String) -> Self {
-            Self { api_key, model }
-        }
-    }
-    #[async_trait]
-    impl LLMProvider for GeminiProvider {
-        fn name(&self) -> &str {
-            &self.model
-        }
-        fn cost_per_1k_tokens(&self) -> f64 {
-            0.0
-        }
-        async fn generate(&self, prompt: &str) -> anyhow::Result<String> {
-            let api_key = std::env::var("GEMINI_API_KEY")
-                .or_else(|_| Ok(self.api_key.clone()))
-                .map_err(|_: std::env::VarError| anyhow::anyhow!("GEMINI_API_KEY not set"))?;
-
-            let client = reqwest::Client::new();
-            let url = format!(
-                "https://generativelanguage.googleapis.com/v1beta/models/{}:generateContent?key={}",
-                self.model, api_key
-            );
-
-            let body = serde_json::json!({
-                "contents": [{"parts": [{"text": prompt}]}],
-                "generationConfig": {
-                    "temperature": 0.7,
-                    "maxOutputTokens": 2048
-                }
-            });
-
-            let response = client
-                .post(&url)
-                .json(&body)
-                .send()
-                .await
-                .map_err(|e| anyhow::anyhow!("Gemini API request failed: {}", e))?;
-
-            let status = response.status();
-            if !status.is_success() {
-                let error_text = response.text().await.unwrap_or_default();
-                tracing::error!("Gemini API error ({}): {}", status, error_text);
-                return Err(anyhow::anyhow!("Gemini API returned error: {}", status));
-            }
-
-            #[derive(Deserialize)]
-            struct GeminiResponse {
-                candidates: Option<Vec<Candidate>>,
-            }
-            #[derive(Deserialize)]
-            struct Candidate {
-                content: Option<Content>,
-            }
-            #[derive(Deserialize)]
-            struct Content {
-                parts: Option<Vec<Part>>,
-            }
-            #[derive(Deserialize)]
-            struct Part {
-                text: Option<String>,
-            }
-
-            let resp: GeminiResponse = response
-                .json()
-                .await
-                .map_err(|e| anyhow::anyhow!("Failed to parse Gemini response: {}", e))?;
-
-            let text = resp
-                .candidates
-                .and_then(|c| c.into_iter().next())
-                .and_then(|c| c.content)
-                .and_then(|c| c.parts)
-                .and_then(|p| p.into_iter().next())
-                .and_then(|p| p.text)
-                .ok_or_else(|| anyhow::anyhow!("No text in Gemini response"))?;
-
-            Ok(text)
-        }
-    }
-
-    #[derive(Debug, Clone)]
-    pub struct GroqProvider {
-        api_key: String,
-        model: String,
-    }
-    impl GroqProvider {
-        pub fn new(api_key: String, model: String) -> Self {
-            Self { api_key, model }
-        }
-    }
-    #[async_trait]
-    impl LLMProvider for GroqProvider {
-        fn name(&self) -> &str {
-            &self.model
-        }
-        fn cost_per_1k_tokens(&self) -> f64 {
-            0.0
-        }
-        async fn generate(&self, prompt: &str) -> anyhow::Result<String> {
-            let api_key = std::env::var("GROQ_API_KEY")
-                .or_else(|_| Ok(self.api_key.clone()))
-                .map_err(|_: std::env::VarError| anyhow::anyhow!("GROQ_API_KEY not set"))?;
-
-            let client = reqwest::Client::new();
-            let url = "https://api.groq.com/openai/v1/chat/completions";
-
-            let body = serde_json::json!({
-                "model": self.model,
-                "messages": [
-                    {"role": "user", "content": prompt}
-                ],
-                "temperature": 0.7,
-                "max_tokens": 2048
-            });
-
-            let response = client
-                .post(url)
-                .header("Authorization", format!("Bearer {}", api_key))
-                .header("Content-Type", "application/json")
-                .json(&body)
-                .send()
-                .await
-                .map_err(|e| anyhow::anyhow!("Groq API request failed: {}", e))?;
-
-            let status = response.status();
-            if !status.is_success() {
-                let error_text = response.text().await.unwrap_or_default();
-                tracing::error!("Groq API error ({}): {}", status, error_text);
-                return Err(anyhow::anyhow!("Groq API returned error: {}", status));
-            }
-
-            #[derive(Deserialize)]
-            struct GroqResponse {
-                choices: Option<Vec<Choice>>,
-            }
-            #[derive(Deserialize)]
-            struct Choice {
-                message: Option<MessageDetail>,
-            }
-            #[derive(Deserialize)]
-            struct MessageDetail {
-                content: Option<String>,
-            }
-
-            let resp: GroqResponse = response
-                .json()
-                .await
-                .map_err(|e| anyhow::anyhow!("Failed to parse Groq response: {}", e))?;
-
-            let text = resp
-                .choices
-                .and_then(|c| c.into_iter().next())
-                .and_then(|c| c.message)
-                .and_then(|m| m.content)
-                .ok_or_else(|| anyhow::anyhow!("No text in Groq response"))?;
-
-            Ok(text)
-        }
-    }
-
-    #[derive(Debug, Clone)]
-    pub struct MinimaxProvider {
-        api_key: String,
-        group_id: String,
-        model: String,
-    }
-    impl MinimaxProvider {
-        pub fn new(api_key: String, group_id: String, model: String) -> Self {
-            Self {
-                api_key,
-                group_id,
-                model,
-            }
-        }
-    }
-    #[async_trait]
-    impl LLMProvider for MinimaxProvider {
-        fn name(&self) -> &str {
-            &self.model
-        }
-        fn cost_per_1k_tokens(&self) -> f64 {
-            0.0
-        }
-        async fn generate(&self, prompt: &str) -> anyhow::Result<String> {
-            let api_key = std::env::var("MINIMAX_API_KEY")
-                .or_else(|_| Ok(self.api_key.clone()))
-                .map_err(|_: std::env::VarError| anyhow::anyhow!("MINIMAX_API_KEY not set"))?;
-
-            let group_id = std::env::var("MINIMAX_GROUP_ID")
-                .or_else(|_| Ok(self.group_id.clone()))
-                .map_err(|_: std::env::VarError| anyhow::anyhow!("MINIMAX_GROUP_ID not set"))?;
-
-            let client = reqwest::Client::new();
-            let url = format!(
-                "https://api.minimax.chat/v1/text/chatcompletion_pro?GroupId={}",
-                group_id
-            );
-
-            let body = serde_json::json!({
-                "model": self.model,
-                "messages": [
-                    {"role": "user", "content": prompt}
-                ],
-                "temperature": 0.7,
-                "max_tokens": 2048
-            });
-
-            let response = client
-                .post(&url)
-                .header("Authorization", format!("Bearer {}", api_key))
-                .header("Content-Type", "application/json")
-                .json(&body)
-                .send()
-                .await
-                .map_err(|e| anyhow::anyhow!("MiniMax API request failed: {}", e))?;
-
-            let status = response.status();
-            if !status.is_success() {
-                let error_text = response.text().await.unwrap_or_default();
-                tracing::error!("MiniMax API error ({}): {}", status, error_text);
-                return Err(anyhow::anyhow!("MiniMax API returned error: {}", status));
-            }
-
-            #[derive(Deserialize)]
-            struct MinimaxResponse {
-                choices: Option<Vec<Choice>>,
-            }
-            #[derive(Deserialize)]
-            struct Choice {
-                messages: Option<Vec<Message>>,
-            }
-            #[derive(Deserialize)]
-            struct Message {
-                text: Option<String>,
-            }
-
-            let resp: MinimaxResponse = response
-                .json()
-                .await
-                .map_err(|e| anyhow::anyhow!("Failed to parse MiniMax response: {}", e))?;
-
-            let text = resp
-                .choices
-                .and_then(|c| c.into_iter().next())
-                .and_then(|c| c.messages)
-                .and_then(|m| m.into_iter().next())
-                .and_then(|m| m.text)
-                .ok_or_else(|| anyhow::anyhow!("No text in MiniMax response"))?;
-
-            Ok(text)
-        }
-    }
-
-    #[derive(Debug, Clone)]
-    pub struct InMemoryCooldownStore;
-    impl Default for InMemoryCooldownStore {
-        fn default() -> Self {
-            Self::new()
-        }
-    }
-    impl InMemoryCooldownStore {
-        pub fn new() -> Self {
-            Self
-        }
-    }
-
-    #[derive(Debug, Clone, Serialize, Deserialize)]
-    pub struct ProviderId {
-        pub provider: String,
-        pub model: String,
-    }
-    impl ProviderId {
-        pub fn new(provider: &str, model: &str) -> Self {
-            Self {
-                provider: provider.to_string(),
-                model: model.to_string(),
-            }
-        }
-    }
-
-    #[derive(Debug, Clone)]
-    pub struct StochasticRotator {
-        providers: Vec<Arc<dyn LLMProvider>>,
-        counter: Arc<std::sync::atomic::AtomicUsize>,
-    }
-    impl StochasticRotator {
-        pub fn new(_store: Arc<InMemoryCooldownStore>) -> Self {
-            // Use time as a simple seed for the initial counter
-            let initial = std::time::SystemTime::now()
-                .duration_since(std::time::UNIX_EPOCH)
-                .map(|d| d.as_secs() as usize)
-                .unwrap_or(0);
-            Self {
-                providers: Vec::new(),
-                counter: Arc::new(std::sync::atomic::AtomicUsize::new(initial)),
-            }
-        }
-        pub fn add_provider(&mut self, _id: ProviderId, provider: Arc<dyn LLMProvider>) {
-            self.providers.push(provider);
-        }
-    }
-    #[async_trait]
-    impl LLMProvider for StochasticRotator {
-        fn name(&self) -> &str {
-            "stochastic-rotator"
-        }
-        fn cost_per_1k_tokens(&self) -> f64 {
-            0.0
-        }
-        async fn generate(&self, prompt: &str) -> anyhow::Result<String> {
-            if self.providers.is_empty() {
-                return Err(anyhow::anyhow!("No providers available"));
-            }
-
-            let start_idx = self
-                .counter
-                .fetch_add(1, std::sync::atomic::Ordering::SeqCst)
-                % self.providers.len();
-
-            let mut last_err = None;
-            for i in 0..self.providers.len() {
-                let idx = (start_idx + i) % self.providers.len();
-                let provider = &self.providers[idx];
-                match provider.generate(prompt).await {
-                    Ok(res) => return Ok(res),
-                    Err(e) => {
-                        last_err = Some(e);
-                        continue;
-                    },
-                }
-            }
-            Err(last_err.unwrap_or_else(|| anyhow::anyhow!("All providers failed")))
-        }
-    }
-
-    #[derive(Debug, Clone, Serialize, Deserialize)]
-    pub struct Decision {
-        pub reasoning: String,
-        pub action: String,
-        pub parameters: Option<Value>,
-        pub confidence: f32,
-        pub providers_used: Vec<String>,
-    }
-
-    #[derive(Debug, Clone)]
-    pub struct DecisionEngine {
-        providers: Vec<Arc<dyn LLMProvider>>,
-    }
-    impl Default for DecisionEngine {
-        fn default() -> Self {
-            Self::new()
-        }
-    }
-    impl DecisionEngine {
-        pub fn new() -> Self {
-            Self {
-                providers: Vec::new(),
-            }
-        }
-        pub fn builder() -> DecisionEngineBuilder {
-            DecisionEngineBuilder {
-                providers: Vec::new(),
-            }
-        }
-        pub fn providers(&self) -> &[Arc<dyn LLMProvider>] {
-            &self.providers
-        }
-        pub async fn decide(&self, ctx: &DecisionContext) -> anyhow::Result<Decision> {
-            // Native resilience: try the first provider (StochasticRotator if configured)
-            // or fallback if needed.
-            if let Some(provider) = self.providers.first() {
-                let prompt = format!(
-                    "QUERY: {}\nSUMMARY: {}\n\nBased on the above, decide the next action.",
-                    ctx.query,
-                    ctx.summary.as_deref().unwrap_or("None")
-                );
-                let _resp = provider.generate(&prompt).await?;
-
-                Ok(Decision {
-                    reasoning: "Resilient decision via synapse-agentic".to_string(),
-                    action: "chat".to_string(),
-                    parameters: None,
-                    confidence: 1.0,
-                    providers_used: vec![provider.name().to_string()],
-                })
-            } else {
-                Ok(Decision {
-                    reasoning: "mock".to_string(),
-                    action: "final answer".to_string(),
-                    parameters: None,
-                    confidence: 1.0,
-                    providers_used: vec!["mock".to_string()],
-                })
-            }
-        }
-    }
-
-    #[derive(Debug, Clone)]
-    pub struct DecisionEngineBuilder {
-        providers: Vec<Arc<dyn LLMProvider>>,
-    }
-    impl DecisionEngineBuilder {
-        pub fn with_provider<P: Provider + 'static>(mut self, p: P) -> Self {
-            self.providers.push(Arc::new(p));
-            self
-        }
-        pub fn build(self) -> DecisionEngine {
-            DecisionEngine {
-                providers: self.providers,
-            }
-        }
-    }
-
-    #[derive(Debug, Clone, Serialize, Deserialize)]
-    pub struct EmptyContext;
-    pub trait ToolContext: Send + Sync {}
-    impl ToolContext for EmptyContext {}
-
-    #[derive(Clone)]
-    pub struct ToolRegistry {
-        tools: Arc<tokio::sync::RwLock<std::collections::HashMap<String, Arc<dyn Tool>>>>,
-    }
-
-    impl std::fmt::Debug for ToolRegistry {
-        fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-            f.debug_struct("ToolRegistry").finish_non_exhaustive()
-        }
-    }
-
-    impl Default for ToolRegistry {
-        fn default() -> Self {
-            Self::new()
-        }
-    }
-    impl ToolRegistry {
-        pub fn new() -> Self {
-            Self {
-                tools: Arc::new(tokio::sync::RwLock::new(std::collections::HashMap::new())),
-            }
-        }
-        pub async fn register_tool<T: Tool + 'static>(&self, tool: T) {
-            let mut tools = self.tools.write().await;
-            tools.insert(tool.name().to_string(), Arc::new(tool));
-        }
-        pub async fn call(
-            &self,
-            name: &str,
-            ctx: &dyn ToolContext,
-            args: Value,
-        ) -> anyhow::Result<Value> {
-            let tools = self.tools.read().await;
-            if let Some(tool) = tools.get(name) {
-                tool.call(ctx, args).await
-            } else {
-                Err(anyhow::anyhow!("Tool not found: {}", name))
-            }
-        }
-    }
-
-    #[async_trait]
-    pub trait Tool: Send + Sync {
-        fn name(&self) -> &str;
-        fn description(&self) -> &str;
-        fn parameters(&self) -> Value;
-        async fn call(&self, ctx: &dyn ToolContext, args: Value) -> anyhow::Result<Value>;
-    }
-
-    #[async_trait]
-    pub trait Agent: Send + Sync + 'static {
-        type Input: Send + 'static;
-        fn name(&self) -> &str;
-        async fn handle(&mut self, msg: Self::Input) -> anyhow::Result<()>;
-    }
-
-    #[derive(Clone)]
-    pub struct AgentHandle<T: Send + 'static> {
-        tx: mpsc::Sender<T>,
-    }
-    impl<T: Send + 'static> AgentHandle<T> {
-        pub async fn send(&self, msg: T) -> Result<(), mpsc::error::SendError<T>> {
-            self.tx.send(msg).await
-        }
-    }
-
-    #[derive(Default)]
-    pub struct Hive;
-    impl Hive {
-        pub fn new() -> Self {
-            Self
-        }
-        pub fn spawn<A>(&mut self, mut agent: A) -> AgentHandle<A::Input>
-        where
-            A: Agent,
-        {
-            let (tx, mut rx) = mpsc::channel::<A::Input>(64);
-            tokio::spawn(async move {
-                while let Some(msg) = rx.recv().await {
-                    let _ = agent.handle(msg).await;
-                }
-            });
-            AgentHandle { tx }
-        }
-    }
-
-    #[derive(Debug, Clone, Serialize, Deserialize)]
-    pub enum MessageRole {
-        System,
-        User,
-        Assistant,
-    }
-
-    #[derive(Debug, Clone, Serialize, Deserialize)]
-    pub struct Message {
-        pub role: MessageRole,
-        pub content: String,
-        pub token_count: Option<u32>,
-    }
-    impl Message {
-        pub fn new(role: MessageRole, content: String) -> Self {
-            Self {
-                role,
-                content,
-                token_count: None,
-            }
-        }
-    }
-
-    #[derive(Debug, Clone)]
-    pub struct MessageChunk {
-        pub messages: Vec<Message>,
-        pub start_index: usize,
-    }
-    impl MessageChunk {
-        pub fn new(messages: Vec<Message>, start_index: usize) -> Self {
-            Self {
-                messages,
-                start_index,
-            }
-        }
-    }
-
-    pub trait TokenCounter: Send + Sync {
-        fn count_tokens(&self, text: &str) -> anyhow::Result<u32>;
-        fn count_message(&self, message: &Message) -> anyhow::Result<u32> {
-            self.count_tokens(&message.content)
-        }
-    }
-
-    #[derive(Debug, Clone)]
-    pub struct SimpleTokenEstimator;
-    impl SimpleTokenEstimator {
-        pub fn new(_model: &str) -> Self {
-            Self
-        }
-    }
-    impl TokenCounter for SimpleTokenEstimator {
-        fn count_tokens(&self, text: &str) -> anyhow::Result<u32> {
-            Ok(text.split_whitespace().count() as u32)
-        }
-    }
-
-    #[derive(Debug, Clone)]
-    pub struct CompactionConfig {
-        pub warning_tokens: u32,
-        pub critical_tokens: u32,
-        pub keep_recent: usize,
-    }
-    impl CompactionConfig {
-        pub fn small_context() -> Self {
-            Self {
-                warning_tokens: 1500,
-                critical_tokens: 2500,
-                keep_recent: 10,
-            }
-        }
-    }
-
-    #[derive(Debug, Clone, Copy, PartialEq, Eq)]
-    pub enum ContextOverflowRisk {
-        Low,
-        Warning,
-        Critical,
-    }
-
-    #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-    pub enum TaskStatus {
-        Pending,
-        InProgress,
-        Completed,
-        Failed,
-    }
-
-    #[derive(Debug, Clone, Serialize, Deserialize)]
-    pub struct PlannedTask {
-        pub id: String,
-        pub description: String,
-        pub estimated_tool: Option<String>,
-        pub status: TaskStatus,
-    }
-
-    #[async_trait]
-    pub trait ExplicitPlanner: Send + Sync {
-        async fn plan(
-            &self,
-            goal: &str,
-            context: &DecisionContext,
-        ) -> anyhow::Result<Vec<PlannedTask>>;
-    }
-
-    #[derive(Debug, Clone)]
-    pub struct SessionContext {
-        cfg: CompactionConfig,
-        messages: Vec<Message>,
-    }
-    impl SessionContext {
-        pub fn new(cfg: CompactionConfig) -> Self {
-            Self {
-                cfg,
-                messages: Vec::new(),
-            }
-        }
-        pub fn add_message(&mut self, msg: Message) {
-            self.messages.push(msg);
-        }
-        pub fn total_tokens(&self) -> u32 {
-            self.messages
-                .iter()
-                .map(|m| m.token_count.unwrap_or(0))
-                .sum()
-        }
-        pub fn overflow_risk(&self) -> ContextOverflowRisk {
-            let total = self.total_tokens();
-            if total >= self.cfg.critical_tokens {
-                ContextOverflowRisk::Critical
-            } else if total >= self.cfg.warning_tokens {
-                ContextOverflowRisk::Warning
-            } else {
-                ContextOverflowRisk::Low
-            }
-        }
-        pub fn compactable_messages(&self) -> &[Message] {
-            if self.messages.len() > self.cfg.keep_recent {
-                &self.messages[..self.messages.len() - self.cfg.keep_recent]
-            } else {
-                &[]
-            }
-        }
-        pub fn recent_messages(&self) -> &[Message] {
-            let keep = self.cfg.keep_recent.min(self.messages.len());
-            &self.messages[self.messages.len().saturating_sub(keep)..]
-        }
-        pub fn recent_messages_mut(&mut self) -> &mut [Message] {
-            let keep = self.cfg.keep_recent.min(self.messages.len());
-            let start = self.messages.len().saturating_sub(keep);
-            &mut self.messages[start..]
-        }
-    }
-
-    #[derive(Debug, Clone, Copy)]
-    pub enum SummarizationStrategy {
-        Technical,
-    }
-
-    #[derive(Debug, Clone)]
-    pub struct LLMSummarizer {
-        provider: Arc<dyn LLMProvider>,
-        _strategy: SummarizationStrategy,
-    }
-    impl LLMSummarizer {
-        pub fn for_technical(provider: Arc<dyn LLMProvider>) -> Self {
-            Self {
-                provider,
-                _strategy: SummarizationStrategy::Technical,
-            }
-        }
-        pub async fn summarize(&self, chunk: &MessageChunk) -> anyhow::Result<Message> {
-            let history = chunk
-                .messages
-                .iter()
-                .map(|m| format!("{:?}: {}", m.role, m.content))
-                .collect::<Vec<_>>()
-                .join("\n");
-            let prompt = format!(
-                "Summarize the following technical conversation history concisely, focusing on actions taken and their outcomes:\n\n{}",
-                history
-            );
-            let summary = self.provider.generate(&prompt).await?;
-            Ok(Message::new(MessageRole::Assistant, summary))
-        }
-    }
+    pub use crate::router::{
+        Agent, AgentHandle, Decision, DecisionEngine, DecisionEngineBuilder, EmptyContext, Hive,
+        InMemoryCooldownStore, ProviderId, StochasticRotator, Tool, ToolContext, ToolRegistry,
+    };
 }
 
 pub mod framework {
     pub mod workflow {
-        use async_trait::async_trait;
-        use serde_json::Value;
-        use std::collections::HashMap;
+        pub use async_trait::async_trait;
+        pub use serde_json::Value;
 
-        #[derive(Debug, Clone)]
-        pub struct ContextState {
-            data: HashMap<String, Value>,
-        }
-
-        impl ContextState {
-            pub fn new(initial: Value) -> Self {
-                let mut data = HashMap::new();
-                if let Value::Object(map) = initial {
-                    for (k, v) in map {
-                        data.insert(k, v);
-                    }
-                }
-                Self { data }
-            }
-
-            pub fn get_string(&self, key: &str) -> Option<String> {
-                self.data
-                    .get(key)
-                    .and_then(|v| v.as_str().map(ToOwned::to_owned))
-            }
-
-            pub fn get_value(&self, key: &str) -> Option<&Value> {
-                self.data.get(key)
-            }
-
-            pub fn set_value(&mut self, key: &str, value: Value) {
-                self.data.insert(key.to_string(), value);
-            }
-        }
-
-        #[derive(Debug, Clone)]
-        pub enum NodeResult {
-            Continue(Option<String>),
-            Error(String),
-            Halt,
-        }
-
-        #[async_trait]
-        pub trait GraphNode: Send + Sync {
-            fn id(&self) -> &str;
-            async fn execute(&mut self, state: &mut ContextState) -> anyhow::Result<NodeResult>;
-        }
-
-        pub struct ReflectionNode {
-            id: String,
-            route_to: String,
-            retries: usize,
-            current: usize,
-        }
-
-        impl ReflectionNode {
-            pub fn new(id: &str, route_to: &str, retries: usize) -> Self {
-                Self {
-                    id: id.to_string(),
-                    route_to: route_to.to_string(),
-                    retries,
-                    current: 0,
-                }
-            }
-        }
-
-        #[async_trait]
-        impl GraphNode for ReflectionNode {
-            fn id(&self) -> &str {
-                &self.id
-            }
-
-            async fn execute(&mut self, _state: &mut ContextState) -> anyhow::Result<NodeResult> {
-                if self.current < self.retries {
-                    self.current += 1;
-                    Ok(NodeResult::Continue(Some(self.route_to.clone())))
-                } else {
-                    Ok(NodeResult::Halt)
-                }
-            }
-        }
-
-        pub struct StateGraph {
-            nodes: HashMap<String, Box<dyn GraphNode>>,
-            entry: Option<String>,
-            error_handler: Option<String>,
-        }
-
-        impl Default for StateGraph {
-            fn default() -> Self {
-                Self::new()
-            }
-        }
-
-        impl StateGraph {
-            pub fn new() -> Self {
-                Self {
-                    nodes: HashMap::new(),
-                    entry: None,
-                    error_handler: None,
-                }
-            }
-
-            pub fn add_node(&mut self, node: Box<dyn GraphNode>) {
-                self.nodes.insert(node.id().to_string(), node);
-            }
-
-            pub fn set_entry_point(&mut self, id: &str) {
-                self.entry = Some(id.to_string());
-            }
-
-            pub fn set_error_handler(&mut self, id: &str) {
-                self.error_handler = Some(id.to_string());
-            }
-
-            pub async fn execute(
-                &mut self,
-                mut state: ContextState,
-            ) -> anyhow::Result<ContextState> {
-                let mut current = self
-                    .entry
-                    .clone()
-                    .ok_or_else(|| anyhow::anyhow!("entry point not configured"))?;
-
-                loop {
-                    let node = self
-                        .nodes
-                        .get_mut(&current)
-                        .ok_or_else(|| anyhow::anyhow!("node '{}' not found", current))?;
-
-                    match node.execute(&mut state).await? {
-                        NodeResult::Halt => break,
-                        NodeResult::Continue(Some(next)) => current = next,
-                        NodeResult::Continue(None) => {},
-                        NodeResult::Error(err) => {
-                            state.set_value("error", Value::String(err));
-                            if let Some(handler) = self.error_handler.clone() {
-                                current = handler;
-                            } else {
-                                break;
-                            }
-                        },
-                    }
-                }
-
-                Ok(state)
-            }
-        }
+        pub use crate::context::ContextState;
+        pub use crate::graph::{GraphNode, NodeResult, ReflectionNode, StateGraph};
     }
 }
